+++ NEWS.md
@@ -9,6 +9,7 @@
 - refine check on whether bias correction is needed in `windRose`
 - still plot data when <6 points in `TheilSen` and when no trend information is given
 - New function `binData` to easily summarise mean and 95% confidence intervals for intervals of a variable
+- Export `bootMeanDF`, used to calculate the bootstrap uncertainty in the mean of a vector.
 
 # openair 2.0-0